Move itemsByFilter to utils
