+++ lib/resources.js
@@ -5,7 +5,6 @@ const ResourceSerializer = require('./jsonld_serializers.js').ResourceSerializer
 const AggregationsSerializer = require('./jsonld_serializers.js').AggregationsSerializer
 const AggregationSerializer = require('./jsonld_serializers.js').AggregationSerializer
 const ItemResultsSerializer = require('./jsonld_serializers.js').ItemResultsSerializer
-const LocationLabelUpdater = require('./location_label_updater')
 const AnnotatedMarcSerializer = require('./annotated-marc-serializer')
 const { makeNyplDataApiClient } = require('./data-api-client')
 const { IndexSearchError, IndexConnectionError } = require('./errors')
@@ -21,7 +20,7 @@ const ElasticQueryBuilder = require('./elasticsearch/elastic-query-builder')
 const { AGGREGATIONS_SPEC, EXCLUDE_FIELDS, SORT_FIELDS } = require('./elasticsearch/config')
 
 const errors = require('./errors')
-const { esRangeValue, parseSearchParams, nyplSourceAndId } = require('./utils/resource-helpers')
+const { esRangeValue, parseSearchParams, nyplSourceAndId, itemsByFilter } = require('./utils/resource-helpers')
 const { bodyForFindByUri, addInnerHits, itemsFilterContext, itemsQueryContext } = require('./elasticsearch/elastic-body-builder')
 
 const RESOURCES_INDEX = process.env.RESOURCES_INDEX
@@ -113,46 +112,6 @@ module.exports = function (app, _private = null) {
       .then(AnnotatedMarcSerializer.serialize)
   }
 
-  function itemsByFilter (filter, opts) {
-    opts = Object.assign({
-      _source: null
-    }, opts)
-
-    // Build ES query body:
-    const body = {
-      query: {
-        nested: {
-          path: 'items',
-          score_mode: 'avg',
-          query: {
-            constant_score: {
-              filter
-            }
-          }
-        }
-      }
-    }
-    if (opts._source) body._source = opts._source
-
-    app.logger.debug('Resources#itemsByFilter', body)
-    return app.esClient.search(body)
-      .then((resp) => {
-        if (!resp || !resp.hits || resp.hits.total === 0) return Promise.reject(new Error('No matching items'))
-        resp = new LocationLabelUpdater(resp).responseWithUpdatedLabels()
-        // Convert this ES bibs response into an array of flattened items:
-        return resp.hits.hits
-          .map((doc) => doc._source)
-          // Reduce to a flat array of items
-          .reduce((a, bib) => {
-            return a.concat(bib.items)
-              // Let's affix that bnum into the item's identifiers so we know where it came from:
-              .map((i) => {
-                return Object.assign(i, { identifier: [`urn:bnum:${bib.uri}`].concat(i.identifier) })
-              })
-          }, [])
-      })
-  }
-
   // Get deliveryLocations for given resource(s)
   app.resources.deliveryLocationsByBarcode = function (params, opts) {
     params = parseParams(params, {
@@ -170,16 +129,7 @@ module.exports = function (app, _private = null) {
       })
 
     // Create promise to resolve items:
-    const fetchItems = itemsByFilter(
-      { terms: { 'items.identifier': identifierValues } },
-      { _source: ['uri', 'type', 'items.uri', 'items.type', 'items.identifier', 'items.holdingLocation', 'items.status', 'items.catalogItemType', 'items.accessMessage', 'items.m2CustomerCode'] }
-
-      // Filter out any items (multi item bib) that don't match one of the queriered barcodes:
-    ).then((items) => {
-      return items.filter((item) => {
-        return item.identifier.filter((i) => identifierValues.indexOf(i) >= 0).length > 0
-      })
-    })
+    const fetchItems = itemsByFilter(identifierValues, app)
 
     // Run both item fetch and patron fetch in parallel:
     return Promise.all([fetchItems, lookupPatronType])
