+++ lib/utils/resource-helpers.js
@@ -2,6 +2,7 @@ const { parseParams } = require('../util')
 const { FILTER_CONFIG, SEARCH_SCOPES, AGGREGATIONS_SPEC, SORT_FIELDS } = require('../elasticsearch/config')
 const NyplSourceMapper = require('research-catalog-indexer/lib/utils/nypl-source-mapper')
 const errors = require('../errors')
+const LocationLabelUpdater = require('../location_label_updater')
 
 /**
  * Given a range represented as an array, returns a corresponding ES range object
@@ -67,8 +68,56 @@ const nyplSourceAndId = async function (params) {
   return { id, nyplSource }
 }
 
+function itemsByFilter (identifierValues, app) {
+  const filter = { terms: { 'items.identifier': identifierValues } }
+  let opts = { _source: ['uri', 'type', 'items.uri', 'items.type', 'items.identifier', 'items.holdingLocation', 'items.status', 'items.catalogItemType', 'items.accessMessage', 'items.m2CustomerCode'] }
+
+  opts = Object.assign({
+    _source: null
+  }, opts)
+
+  // Build ES query body:
+  const body = {
+    query: {
+      nested: {
+        path: 'items',
+        score_mode: 'avg',
+        query: {
+          constant_score: {
+            filter
+          }
+        }
+      }
+    }
+  }
+  if (opts._source) body._source = opts._source
+
+  app.logger.debug('Resources#itemsByFilter', body)
+  return app.esClient.search(body)
+    .then((resp) => {
+      if (!resp || !resp.hits || resp.hits.total === 0) return Promise.reject(new Error('No matching items'))
+      resp = new LocationLabelUpdater(resp).responseWithUpdatedLabels()
+      // Convert this ES bibs response into an array of flattened items:
+      return resp.hits.hits
+        .map((doc) => doc._source)
+        // Reduce to a flat array of items
+        .reduce((a, bib) => {
+          return a.concat(bib.items)
+            // Let's affix that bnum into the item's identifiers so we know where it came from:
+            .map((i) => {
+              return Object.assign(i, { identifier: [`urn:bnum:${bib.uri}`].concat(i.identifier) })
+            })
+        }, [])
+    }).then((items) => {
+      return items.filter((item) => {
+        return item.identifier.filter((i) => identifierValues.indexOf(i) >= 0).length > 0
+      })
+    })
+}
+
 module.exports = {
   esRangeValue,
   parseSearchParams,
-  nyplSourceAndId
+  nyplSourceAndId,
+  itemsByFilter
 }
